requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.128.0
uvicorn==0.40.0
langchain-chroma==1.1.0
langchain-core==1.2.7
langchain-huggingface==1.2.0
sentence-transformers==5.2.0
huggingface-hub>=0.33.4,<1.0.0
torch==2.9.1
python-dotenv==1.2.1
requests==2.32.5
numpy==2.4.1
Jinja2==3.1.6
MarkupSafe==3.0.3
diskcache==5.6.3
typing_extensions==4.15.0
pydantic==2.12.5
pydantic_core==2.41.5
chromadb==1.4.0